Add marginTop/Bottom to DottedLine and refactor PopupRecord
