+++ src/components/Mypage/PopupRecord/ReportOption.tsx
@@ -1,0 +1,49 @@
+import React from "react";
+import { theme } from "styles/theme";
+
+export type ReportOptionType = "Reporting" | "Reported";
+
+type ReportOptionProps = {
+  option: ReportOptionType;
+  onClick: (option: ReportOptionType) => void;
+};
+
+const ReportOption = (props: ReportOptionProps) => {
+  const styleContainer = {
+    display: "flex",
+    columnGap: "8px",
+    padding: "10px 0 12px 8px",
+  };
+  const styleButton = (isSelected: boolean) => {
+    return {
+      display: "flex",
+      ...theme.font10,
+      borderRadius: "4px",
+      padding: "3px 6px 3px",
+      cursor: theme.cursor(),
+      color: isSelected ? theme.white : theme.gray_text,
+      backgroundColor: isSelected ? theme.purple : theme.gray_background,
+      boxShadow: isSelected
+        ? theme.shadow_purple_button_inset
+        : theme.shadow_gray_input_inset,
+    };
+  };
+  return (
+    <div style={styleContainer}>
+      <div
+        style={styleButton(props.option === "Reporting")}
+        onClick={() => props.onClick("Reporting")}
+      >
+        신고한 내역
+      </div>
+      <div
+        style={styleButton(props.option === "Reported")}
+        onClick={() => props.onClick("Reported")}
+      >
+        신고 받은 내역
+      </div>
+    </div>
+  );
+};
+
+export default ReportOption;
